Run flake8 tests on all new code submissions
The Travis CI services is free for all open source projects like this one.  The owner of the this repo would need to go to https://travis-ci.org/profile (log in via GitHub) and flip the repository switch __on__ to enable free automated flake8 testing of each pull request.  We are testing both Python 2 and Python 3 in parallel but Python 3 has been set to allow_failures until the port is complete.